+++ src/main/java/dinkplugin/notifiers/data/DiaryNotificationData.java
@@ -1,11 +1,11 @@
 package dinkplugin.notifiers.data;
 
-import dinkplugin.domain.AchievementDiaries;
+import dinkplugin.domain.AchievementDiary;
 import lombok.Value;
 
 @Value
 public class DiaryNotificationData {
     String area;
-    AchievementDiaries.Difficulty difficulty;
+    AchievementDiary.Difficulty difficulty;
     int total;
 }