+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Minor: Use notifier-specific screenshot filenames. (#79)
 - Minor: Support notifier-specific webhook URLs. (#78)
+- Dev: Explicitly define achievement diaries. (#80)
 - Dev: Add mockito test suite for notifiers. (#74, #75)
 - Dev: Utilize more dependency injection. (#73)